[add] support for FT.SUGLEN, FT.SUGDEL. Included default options for FT.SUGGET. Increased coverage
