+++ redisearch/client.go
@@ -12,31 +12,6 @@ import (
 	"github.com/garyburd/redigo/redis"
 )
 
-// Options are flags passed to the the abstract Index call, which receives them as interface{}, allowing
-// for implementation specific options
-type Options struct {
-
-	// If set, we will not save the documents contents, just index them, for fetching ids only
-	NoSave bool
-
-	NoFieldFlags bool
-
-	NoFrequencies bool
-
-	NoOffsetVectors bool
-
-	Stopwords []string
-}
-
-// DefaultOptions represents the default options
-var DefaultOptions = Options{
-	NoSave:          false,
-	NoFieldFlags:    false,
-	NoFrequencies:   false,
-	NoOffsetVectors: false,
-	Stopwords:       nil,
-}
-
 // Client is an interface to redisearch's redis commands
 type Client struct {
 	pool ConnPool
@@ -66,191 +41,18 @@ func NewClient(addr, name string) *Client {
 }
 
 // CreateIndex configues the index and creates it on redis
-func (i *Client) CreateIndex(s *Schema) error {
+func (i *Client) CreateIndex(s *Schema) (err error) {
 	args := redis.Args{i.name}
 	// Set flags based on options
-	if s.Options.NoFieldFlags {
-		args = append(args, "NOFIELDS")
-	}
-	if s.Options.NoFrequencies {
-		args = append(args, "NOFREQS")
-	}
-	if s.Options.NoOffsetVectors {
-		args = append(args, "NOOFFSETS")
-	}
-	if s.Options.Stopwords != nil {
-		args = args.Add("STOPWORDS", len(s.Options.Stopwords))
-		if len(s.Options.Stopwords) > 0 {
-			args = args.AddFlat(s.Options.Stopwords)
-		}
-	}
-
-	args = append(args, "SCHEMA")
-	for _, f := range s.Fields {
-
-		switch f.Type {
-		case TextField:
-
-			args = append(args, f.Name, "TEXT")
-			if f.Options != nil {
-				opts, ok := f.Options.(TextFieldOptions)
-				if !ok {
-					return errors.New("Invalid text field options type")
-				}
-
-				if opts.Weight != 0 && opts.Weight != 1 {
-					args = append(args, "WEIGHT", opts.Weight)
-				}
-				if opts.NoStem {
-					args = append(args, "NOSTEM")
-				}
-
-				if opts.Sortable {
-					args = append(args, "SORTABLE")
-				}
-
-				if opts.NoIndex {
-					args = append(args, "NOINDEX")
-				}
-			}
-
-		case NumericField:
-			args = append(args, f.Name, "NUMERIC")
-			if f.Options != nil {
-				opts, ok := f.Options.(NumericFieldOptions)
-				if !ok {
-					return errors.New("Invalid numeric field options type")
-				}
-
-				if opts.Sortable {
-					args = append(args, "SORTABLE")
-				}
-				if opts.NoIndex {
-					args = append(args, "NOINDEX")
-				}
-			}
-		case TagField:
-			args = append(args, f.Name, "TAG")
-			if f.Options != nil {
-				opts, ok := f.Options.(TagFieldOptions)
-				if !ok {
-					return errors.New("Invalid tag field options type")
-				}
-				if opts.Separator != 0 {
-					args = append(args, "SEPARATOR", fmt.Sprintf("%c", opts.Separator))
-
-				}
-				if opts.Sortable {
-					args = append(args, "SORTABLE")
-				}
-				if opts.NoIndex {
-					args = append(args, "NOINDEX")
-				}
-			}
-		default:
-			return fmt.Errorf("Unsupported field type %v", f.Type)
-		}
-
+	args, err = SerializeSchema(s, args)
+	if err != nil {
+		return
 	}
 
 	conn := i.pool.Get()
 	defer conn.Close()
-	_, err := conn.Do("FT.CREATE", args...)
+	_, err = conn.Do("FT.CREATE", args...)
 	return err
-}
-
-// IndexingOptions represent the options for indexing a single document
-type IndexingOptions struct {
-	Language         string
-	NoSave           bool
-	Replace          bool
-	Partial          bool
-	ReplaceCondition string
-}
-
-// DefaultIndexingOptions are the default options for document indexing
-var DefaultIndexingOptions = IndexingOptions{
-	Language:         "",
-	NoSave:           false,
-	Replace:          false,
-	Partial:          false,
-	ReplaceCondition: "",
-}
-
-// IndexOptions indexes multiple documents on the index, with optional Options passed to options
-func (i *Client) IndexOptions(opts IndexingOptions, docs ...Document) error {
-
-	conn := i.pool.Get()
-	defer conn.Close()
-
-	n := 0
-	var merr MultiError
-
-	for ii, doc := range docs {
-		args := make(redis.Args, 0, 6+len(doc.Properties))
-		args = append(args, i.name, doc.Id, doc.Score)
-		// apply options
-		if opts.NoSave {
-			args = append(args, "NOSAVE")
-		}
-		if opts.Language != "" {
-			args = append(args, "LANGUAGE", opts.Language)
-		}
-
-		if opts.Partial {
-			opts.Replace = true
-		}
-
-		if opts.Replace {
-			args = append(args, "REPLACE")
-			if opts.Partial {
-				args = append(args, "PARTIAL")
-			}
-			if opts.ReplaceCondition != "" {
-				args = append(args, "IF", opts.ReplaceCondition)
-			}
-		}
-
-		if doc.Payload != nil {
-			args = args.Add("PAYLOAD", doc.Payload)
-		}
-
-		args = append(args, "FIELDS")
-
-		for k, f := range doc.Properties {
-			args = append(args, k, f)
-		}
-
-		if err := conn.Send("FT.ADD", args...); err != nil {
-			if merr == nil {
-				merr = NewMultiError(len(docs))
-			}
-			merr[ii] = err
-
-			return merr
-		}
-		n++
-	}
-
-	if err := conn.Flush(); err != nil {
-		return err
-	}
-
-	for n > 0 {
-		if _, err := conn.Receive(); err != nil {
-			if merr == nil {
-				merr = NewMultiError(len(docs))
-			}
-			merr[n-1] = err
-		}
-		n--
-	}
-
-	if merr == nil {
-		return nil
-	}
-
-	return merr
 }
 
 // convert the result from a redis query to a proper Document object
@@ -472,24 +274,6 @@ func (i *Client) Delete(docId string, deleteDocument bool) (err error) {
 	return
 }
 
-// IndexInfo - Structure showing information about an existing index
-type IndexInfo struct {
-	Schema               Schema
-	Name                 string  `redis:"index_name"`
-	DocCount             uint64  `redis:"num_docs"`
-	RecordCount          uint64  `redis:"num_records"`
-	TermCount            uint64  `redis:"num_terms"`
-	MaxDocID             uint64  `redis:"max_doc_id"`
-	InvertedIndexSizeMB  float64 `redis:"inverted_sz_mb"`
-	OffsetVectorSizeMB   float64 `redis:"offset_vector_sz_mb"`
-	DocTableSizeMB       float64 `redis:"doc_table_size_mb"`
-	KeyTableSizeMB       float64 `redis:"key_table_size_mb"`
-	RecordsPerDocAvg     float64 `redis:"records_per_doc_avg"`
-	BytesPerRecordAvg    float64 `redis:"bytes_per_record_avg"`
-	OffsetsPerTermAvg    float64 `redis:"offsets_per_term_avg"`
-	OffsetBitsPerTermAvg float64 `redis:"offset_bits_per_record_avg"`
-}
-
 func (info *IndexInfo) setTarget(key string, value interface{}) error {
 	v := reflect.ValueOf(info).Elem()
 	for i := 0; i < v.NumField(); i++ {
@@ -637,4 +421,120 @@ func (i *Client) Info() (*IndexInfo, error) {
 	}
 
 	return &ret, nil
+}
+
+// Delete deletes the Autocompleter key for this AC
+func (i *Client) AutocompleterDelete(a *Autocompleter) error {
+	conn := i.pool.Get()
+	defer conn.Close()
+
+	_, err := conn.Do("DEL", a.name)
+	return err
+}
+
+// AddTerms pushes new term suggestions to the index
+func (c *Client) AutocompleterAddTerms(a *Autocompleter, terms ...Suggestion) error {
+	conn := c.pool.Get()
+	defer conn.Close()
+
+	i := 0
+	for _, term := range terms {
+
+		args := redis.Args{a.name, term.Term, term.Score}
+		if term.Payload != "" {
+			args = append(args, "PAYLOAD", term.Payload)
+		}
+
+		if err := conn.Send("FT.SUGADD", args...); err != nil {
+			return err
+		}
+		i++
+	}
+	if err := conn.Flush(); err != nil {
+		return err
+	}
+	for i > 0 {
+		if _, err := conn.Receive(); err != nil {
+			return err
+		}
+		i--
+	}
+	return nil
+}
+
+// AddTerms pushes new term suggestions to the index
+func (c *Client) AutocompleterDeleteTerms(a *Autocompleter, terms ...Suggestion) error {
+	conn := c.pool.Get()
+	defer conn.Close()
+
+	i := 0
+	for _, term := range terms {
+
+		args := redis.Args{a.name, term.Term}
+		if err := conn.Send("FT.SUGDEL", args...); err != nil {
+			return err
+		}
+		i++
+	}
+	if err := conn.Flush(); err != nil {
+		return err
+	}
+	for i > 0 {
+		if _, err := conn.Receive(); err != nil {
+			return err
+		}
+		i--
+	}
+	return nil
+}
+
+// AddTerms pushes new term suggestions to the index
+func (c *Client) AutocompleterLength(a *Autocompleter) (len int64, err error) {
+	conn := c.pool.Get()
+	defer conn.Close()
+	len, err = redis.Int64(conn.Do("FT.SUGLEN", a.name))
+	return
+}
+
+// Suggest gets completion suggestions from the Autocompleter dictionary to the given prefix.
+// If fuzzy is set, we also complete for prefixes that are in 1 Levenshten distance from the
+// given prefix
+//
+// Deprecated: Please use SuggestOpts() instead
+func (c *Client) AutocompleterSuggest(a *Autocompleter, prefix string, num int, fuzzy bool) (ret []Suggestion, err error) {
+	conn := c.pool.Get()
+	defer conn.Close()
+
+	seropts := DefaultSuggestOptions
+	seropts.Num = num
+	seropts.Fuzzy = fuzzy
+	args, inc := a.Serialize(prefix, seropts)
+
+	vals, err := redis.Strings(conn.Do("FT.SUGGET", args...))
+	if err != nil {
+		return nil, err
+	}
+
+	ret = ProcessSugGetVals(vals, inc, true, false)
+
+	return
+}
+
+// SuggestOpts gets completion suggestions from the Autocompleter dictionary to the given prefix.
+// SuggestOptions are passed allowing you specify if the returned values contain a payload, and scores.
+// If SuggestOptions.Fuzzy is set, we also complete for prefixes that are in 1 Levenshten distance from the
+// given prefix
+func (c *Client) AutocompleterSuggestOpts(a *Autocompleter, prefix string, opts SuggestOptions) (ret []Suggestion, err error) {
+	conn := c.pool.Get()
+	defer conn.Close()
+
+	args, inc := a.Serialize(prefix, opts)
+	vals, err := redis.Strings(conn.Do("FT.SUGGET", args...))
+	if err != nil {
+		return nil, err
+	}
+
+	ret = ProcessSugGetVals(vals, inc, opts.WithScores, opts.WithPayloads)
+
+	return
 }
